requirements.txt
torch
torchvision
ultralytics
albumentations
opencv-python
matplotlib
scikit-learn
numpy
pillow
